Started working on 'userProperties' page
Paginated results, but did not filter them according to the logged in
userId selection criteria (meaning that the properties displayed are all
the properties in the database). The delete button triggers a
window.alert with a confirmation function. The function must be updated
in order for it to call the deleteProperty(propertyId) method. The
properties in the database must have a userId associated to them.